refactor(earn): move provider/liquidity label helpers into earn-format
React Doctor: non-component exports in a component file break Fast
Refresh; earn-format.ts is the existing pure-formatter module.
