+++ apps/sdp-web/src/app/dashboard/markets/treasury-solutions/treasury-solutions-workspace.tsx
@@ -47,14 +47,12 @@ import {
   type EarnFundingWallet,
   useEarnFundingWallets,
 } from "../earn/deposit/earn-funding-wallets";
-import { formatUsd } from "../earn/earn-format";
+import { earnProviderLabel, earnStrategyLiquidityLabel, formatUsd } from "../earn/earn-format";
 import {
   EarnDepositAvailabilityBadge,
   EarnStrategyIdentity,
   earnMintAsset,
-  earnProviderLabel,
   earnStrategyAsset,
-  earnStrategyLiquidityLabel,
   earnStrategyReferenceKey,
   formatProviderAmount,
   formatProviderApy,
